move xds endpoint parsing into the xds module, actually parse addresses
